Update MovementHandler.java
I was recently looking at related files in the autonomous section. With my current knowledge of java, i know that for any java application, there is always a main method involved, which is the first two lines of code that I entered in the program. Before, you were not using this main method. Is it unnecessary in this case? If so, why? how were you working without this? thanks,